ai-python/requirements.txt
# FastAPI Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Environment
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.26.0

# Validation
pydantic>=2.5.3
